requirements.txt
fastapi
pydantic
google-genai
pinecone
python-dotenv